Provide an easy way to run collectstatic in the virtualenv
This was a great opportunity to tidy up our scripts-to-rule-them-all,
removing the duplication of the virtualenv handling code, and making
them all work regardless of the script’s location relative to the
current working directory.

We use `$(dirname "$0")` to work out the path to `/scripts` relative
to the user’s current working directory. There are some cases when
this might not work (mostly involving symlinks), but in our intended
use case (where the scripts are at `/vagrant/bluetail/scripts/`)
that shouldn’t be an issue, and the use of `dirname` rather than
various `BASH_SOURCE` / `readlink` alternatives at least means the
scripts should be fairly portable across shells and platforms.